grpc: initial changes for networking protocol
Add a declarative interface for network interfaces.  The protocol
has been defined to include "AddInterface," "UpdateInterface" and
"RemoveInterface."  Under most use cases, update interface should be the
only required call, as link device should exist if being passed in by
the hypervisor and it should only be updated with the desired values.
In this scenario, the interface is identified by the MAC address.

RemoveInterface could be used to delete a link before doing a hot-unplug
of the network device.  In the event of failure for update and delete,
the current state of the interface is returned.

Messages for Add, Update and Remove route are available, though today
only Add and Remove are handled. There have been issues with existing
handling of routes in Clear Containers, and we're looking to resolve these before
moving to an API that just uses UpdateRoutes Kata.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>